Fix create recipe form inconsistency
reuse same recipe interface and ui component for create and edit page
